pyproject.toml
[project]
name = "lobster-claws"
version = "0.1.0"
requires-python = ">=3.11"

[tool.uv]
package = false

[tool.uv.workspace]
members = ["common", "cli", "skills/*", "servers/*"]

[dependency-groups]
dev = ["claws-calendar", "claws-cli", "claws-common", "claws-drive", "claws-gmail", "claws-tasks", "claws-transcribe", "whisper-server", "google-auth-server", "ruff>=0.15", "pytest>=9.0", "pytest-httpx>=0.36"]

[tool.uv.sources]
claws-cli = { workspace = true }
claws-common = { workspace = true }
claws-transcribe = { workspace = true }
whisper-server = { workspace = true }
claws-calendar = { workspace = true }
claws-drive = { workspace = true }
claws-gmail = { workspace = true }
claws-tasks = { workspace = true }
google-auth-server = { workspace = true }

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["cli/tests", "common/tests", "skills", "servers", "tests"]
addopts = "--import-mode=importlib"
